:truck: put `PageConfig` in its own module
